requirements.txt
streamlit
openai-whisper
transformers
torch
pydub
SpeechRecognition
pandas